Add initial functionality
Core attack module included.  Included works mostly as modules with
limited native script functionality.  Basic ARP poisoner is not core
to the project but is included for those that need or want it.  The
selenium PoC shows off functionality as a standalone module.  Config
module is essential to working with all modules now and for those to
come later.